docs: changes swagger config
