+++ src/globals/constants.ts
@@ -1,4 +1,5 @@
 export const appName = 'LoryBlu';
+export const appDescription = process.env.npm_package_description;
 export const appVersion = process.env.npm_package_version;
 export const appLicense = process.env.npm_package_license;
